Update Ban Appeal addon to 1.0.1
Update Ban Appeal addon, no longer require User class to be modified,
make use of the new custom mod sidebar link feature
